feat: add refresh token
Since the token expires every 1 hour, a method that refreshes the user's token should be created to keep it logged in on the application

This commit adds a new endpoint to the Auth module - /refresh, which is responsible for handeling requests when the token is expired and refreshing it again. It is also documented with JSDoc and Swagger.