feat: add relation Many-to-Many between User and Class modules
Since several different teachers can teach several different classes, the need for a Many-to-Many relationship arose.

This commit adds a new Assign Module that has the function of assigning a user to a class, and unassigning them as well. It also changed the findOne and findAll functions from the userService and classService to show the users/classes id's that are assigned to them.